Fix priority of sale_dynamic_price views
Because the views replace the field standard_price, this may cause
conflicts with other modules such as stock_account.

stock_account also replaces the field to add extra behavior.

Decreasing the priority on the views prevents such conflicts
because the field standard_price is moved before other changes are applied.