Add logistics for rented products
Automatically define a rental route for each warehouse.

When a sale order of type rental is confirmed, the rental route
of the warehouse is used.

Add propagation of rental dates.

Propagate the dates from the rental service line to components.
Propagate the dates from the components to the stock moves.