TA#35418 allow to send quotation on payment transaction
Improve the module payment_auto_confirm_sale_order to allow
sending the quotation instead of confirming the order.

Confirming the order is useful for orders that are paid at the end of months (for example).

Sending the quotation is useful for credit cards. It prevents the sale order to remain in draft
state while the transaction is being processed.