TA#43464 set high sale order line sequence
In vanilla Odoo, the default value is 10.

Without this code fix, adding a new component to an existing kit will not work
if the sale order has more than 10 lines.

When adding a new line, a value of 10 does not reflect properly the placement
of the line in the list view. It results in an unexpected permutation of the lines
when the sequences are recomputed.